requirements.txt
ezdxf==1.3.1
maf==0.0.0
matplotlib==3.8.0
numpy==1.26.4
open3d==0.18.0
pandas==2.2.2
paramiko==3.4.0
Pillow==10.3.0
scikit_learn==1.2.2
scipy==1.11.4
setuptools==68.2.2
skorch==1.0.0
torch==1.12.0+cu113
torchvision==0.13.0
trimesh==4.4.0
